rotating verb section
One thing to note is that the 'heading' is not being used from the json, it has been hardcoded into the component. This provides more flexibilty on where to put the line breaks and what word to replace, but if we want this component to be more reusable, we could think of a way to improve this.